accessible/tests/mochitest/test_textattrs.html

<html>

<head>
  <title>Text attributes tests</title>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/common.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/attributes.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/events.js"></script>

  <script type="application/javascript">

    const nsIDOMNSEditableElement =
      Components.interfaces.nsIDOMNSEditableElement;

    var gComputedStyle = null;

    var gTextAttrChangedEventHandler = {
      handleEvent: function gTextAttrChangedEventHandler_handleEvent(aEvent)
      {
        this.eventNumber++;
      },

      eventNumber: 0
    };

    function testSpellTextAttrs()
    {
      registerA11yEventListener(nsIAccessibleEvent.EVENT_TEXT_ATTRIBUTE_CHANGED,
                                gTextAttrChangedEventHandler);

      ID = "area8";
  
      var node = document.getElementById(ID);
      node.setAttribute("value", "valid text inalid tixt");
      node.focus();

      var editor = node.QueryInterface(nsIDOMNSEditableElement).editor;
      var spellchecker = editor.getInlineSpellChecker(true);
      spellchecker.enableRealTimeSpell = true;

      window.setTimeout(function()
        {
          gComputedStyle = document.defaultView.getComputedStyle(node, "");
          var defAttrs = {
            "font-style": gComputedStyle.fontStyle,
            "font-size": (MAC) ? "8pt" : "10pt",
            "background-color": gComputedStyle.backgroundColor,
            "font-weight": gComputedStyle.fontWeight,
            "color": gComputedStyle.color,
            "font-family": gComputedStyle.fontFamily,
            "text-position": gComputedStyle.verticalAlign
          };
          testDefaultTextAttrs(ID, defAttrs);

          var attrs = { };
          var misspelledAttrs = {
            "invalid": "spelling"
          };

          testTextAttrs(ID, 0, attrs, defAttrs, 0, 11);
          testTextAttrs(ID, 11, misspelledAttrs, defAttrs, 11, 17);
          testTextAttrs(ID, 17, attrs, defAttrs, 17, 18);
          testTextAttrs(ID, 18, misspelledAttrs, defAttrs, 18, 22);

          is(gTextAttrChangedEventHandler.eventNumber, 2,
             "Wrong count of 'text attribute changed' events for " + ID);

          unregisterA11yEventListener(nsIAccessibleEvent.EVENT_TEXT_ATTRIBUTE_CHANGED,
                                      gTextAttrChangedEventHandler);

          SimpleTest.finish();
        }, 0);
    }

    function doTest()
    {
      //////////////////////////////////////////////////////////////////////////
      // area1
      var ID = "area1";
      var tempElem = document.getElementById(ID);
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      var defAttrs = {
        "font-style": gComputedStyle.fontStyle,
        "font-size": "10pt",
        "background-color": "rgb(255, 255, 255)",
        "font-weight": gComputedStyle.fontWeight,
        "color": gComputedStyle.color,
        "font-family": gComputedStyle.fontFamily,
        "text-position": gComputedStyle.verticalAlign
      };

      testDefaultTextAttrs(ID, defAttrs);

      var attrs = {};
      testTextAttrs(ID, 0, attrs, defAttrs, 0, 7);

      tempElem = tempElem.firstChild.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"font-weight": gComputedStyle.fontWeight};
      testTextAttrs(ID, 7, attrs, defAttrs, 7, 11);

      attrs = {};
      testTextAttrs(ID, 12, attrs, defAttrs, 11, 18);

      //////////////////////////////////////////////////////////////////////////
      // area2
      ID = "area2";
      tempElem = document.getElementById(ID);
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      defAttrs = {
        "font-style": gComputedStyle.fontStyle,
        "font-size": "14pt",
        "background-color": "rgb(255, 255, 255)",
        "font-weight": gComputedStyle.fontWeight,
        "color": gComputedStyle.color,
        "font-family": gComputedStyle.fontFamily,
        "text-position": gComputedStyle.verticalAlign
      };

      testDefaultTextAttrs(ID, defAttrs);

      attrs = {};
      testTextAttrs(ID, 0, attrs, defAttrs, 0, 7);

      tempElem = tempElem.firstChild.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"font-weight": gComputedStyle.fontWeight};
      testTextAttrs(ID, 7, attrs, defAttrs, 7, 12);

      tempElem = tempElem.firstChild.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"font-style": gComputedStyle.fontStyle,
               "font-weight": gComputedStyle.fontWeight};
      testTextAttrs(ID, 13, attrs, defAttrs, 12, 19);

      tempElem = tempElem.parentNode;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"font-weight": "401"};
      testTextAttrs(ID, 20, attrs, defAttrs, 19, 23);

      attrs = {};
      testTextAttrs(ID, 24, attrs, defAttrs, 23, 30);

      //////////////////////////////////////////////////////////////////////////
      // area3
      ID = "area3";
      tempElem = document.getElementById(ID);
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      defAttrs = {
        "font-style": gComputedStyle.fontStyle,
        "font-size": "12pt",
        "background-color": gComputedStyle.backgroundColor,
        "font-weight": gComputedStyle.fontWeight,
        "color": gComputedStyle.color,
        "font-family": gComputedStyle.fontFamily,
        "text-position": gComputedStyle.verticalAlign
      };

      testDefaultTextAttrs(ID, defAttrs);

      tempElem = tempElem.firstChild.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"color": gComputedStyle.color};
      testTextAttrs(ID, 0, attrs, defAttrs, 0, 6);

      tempElem = tempElem.firstChild.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"color": gComputedStyle.color};
      testTextAttrs(ID, 6, attrs, defAttrs, 6, 26);

      tempElem = tempElem.parentNode;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"color": gComputedStyle.color};
      testTextAttrs(ID, 26, attrs, defAttrs, 26, 27);

      tempElem = tempElem.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"color": gComputedStyle.color,
               "background-color": gComputedStyle.backgroundColor};
      testTextAttrs(ID, 27, attrs, defAttrs, 27, 50);

      //////////////////////////////////////////////////////////////////////////
      // area4
      ID = "area4";
      tempElem = document.getElementById(ID);
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      defAttrs = {
        "font-style": gComputedStyle.fontStyle,
        "font-size": "12pt",
        "background-color": "rgb(255, 255, 255)",
        "font-weight": gComputedStyle.fontWeight,
        "color": gComputedStyle.color,
        "font-family": gComputedStyle.fontFamily,
        "text-position": gComputedStyle.verticalAlign
      };

      testDefaultTextAttrs(ID, defAttrs);

      tempElem = tempElem.firstChild.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"color": gComputedStyle.color};
      testTextAttrs(ID, 0, attrs, defAttrs, 0, 16);

      tempElem = tempElem.nextSibling.firstChild.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"color": gComputedStyle.color};
      testTextAttrs(ID, 16, attrs, defAttrs, 16, 33);

      tempElem = tempElem.parentNode;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"color": gComputedStyle.color};
      testTextAttrs(ID, 34, attrs, defAttrs, 33, 46);

      //////////////////////////////////////////////////////////////////////////
      // area5
      ID = "area5";
      tempElem = document.getElementById(ID);
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      defAttrs = {
        "font-style": gComputedStyle.fontStyle,
        "font-size": "12pt",
        "background-color": "rgb(255, 255, 255)",
        "font-weight": gComputedStyle.fontWeight,
        "color": gComputedStyle.color,
        "font-family": gComputedStyle.fontFamily,
        "text-position": gComputedStyle.verticalAlign
      };

      testDefaultTextAttrs(ID, defAttrs);

      tempElem = tempElem.firstChild.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"color": gComputedStyle.color};
      testTextAttrs(ID, 0, attrs, defAttrs, 0, 5);

      attrs = {};
      testTextAttrs(ID, 7, attrs, defAttrs, 5, 8);

      tempElem = tempElem.nextSibling.nextSibling.nextSibling.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"color": gComputedStyle.color};
      testTextAttrs(ID, 9, attrs, defAttrs, 8, 11);

      attrs = {};
      testTextAttrs(ID, 11, attrs, defAttrs, 11, 18);

      //////////////////////////////////////////////////////////////////////////
      // area6 (CSS vertical-align property, bug 445938)
      ID = "area6";
      tempElem = document.getElementById(ID);
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      defAttrs = {
        "font-style": gComputedStyle.fontStyle,
        "font-size": "12pt",
        "background-color": "rgb(255, 255, 255)",
        "font-weight": gComputedStyle.fontWeight,
        "color": gComputedStyle.color,
        "font-family": gComputedStyle.fontFamily,
        "text-position": gComputedStyle.verticalAlign
      };

      testDefaultTextAttrs(ID, defAttrs);

      attrs = {};
      testTextAttrs(ID, 0, attrs, defAttrs, 0, 5);

      tempElem = tempElem.firstChild.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"text-position": gComputedStyle.verticalAlign,
               "font-size": "10pt"};
      testTextAttrs(ID, 5, attrs, defAttrs, 5, 13);

      attrs = {};
      testTextAttrs(ID, 13, attrs, defAttrs, 13, 27);

      tempElem = tempElem.nextSibling.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"text-position": gComputedStyle.verticalAlign};
      testTextAttrs(ID, 27, attrs, defAttrs, 27, 35);

      attrs = {};
      testTextAttrs(ID, 35, attrs, defAttrs, 35, 39);

      tempElem = tempElem.nextSibling.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"text-position": gComputedStyle.verticalAlign,
               "font-size": "10pt"};
      testTextAttrs(ID, 39, attrs, defAttrs, 39, 50);

      attrs = {};
      testTextAttrs(ID, 50, attrs, defAttrs, 50, 55);

      tempElem = tempElem.nextSibling.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"text-position": gComputedStyle.verticalAlign};
      testTextAttrs(ID, 55, attrs, defAttrs, 55, 64);

      //////////////////////////////////////////////////////////////////////////
      // area7
      ID = "area7";
      tempElem = document.getElementById(ID);
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      defAttrs = {
        "language": "en",
        "font-style": gComputedStyle.fontStyle,
        "font-size": "12pt",
        "background-color": "rgb(255, 255, 255)",
        "font-weight": gComputedStyle.fontWeight,
        "color": gComputedStyle.color,
        "font-family": gComputedStyle.fontFamily,
        "text-position": gComputedStyle.verticalAlign
      };

      testDefaultTextAttrs(ID, defAttrs);

      attrs = {"language": "ru"};
      testTextAttrs(ID, 0, attrs, defAttrs, 0, 12);

      attrs = {};
      testTextAttrs(ID, 12, attrs, defAttrs, 12, 13);

      tempElem = tempElem.firstChild.nextSibling.nextSibling.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = { "background-color": gComputedStyle.backgroundColor};
      testTextAttrs(ID, 13, attrs, defAttrs, 13, 26);

      attrs = {};
      testTextAttrs(ID, 26, attrs, defAttrs, 26, 27);

      attrs = {"language": "de"};
      testTextAttrs(ID, 27, attrs, defAttrs, 27, 42);

      attrs = {};
      testTextAttrs(ID, 42, attrs, defAttrs, 42, 50);

      attrs = {};
      testTextAttrs(ID, 43, attrs, defAttrs, 42, 50);

      tempElem = tempElem.nextSibling.nextSibling.nextSibling.nextSibling.firstChild.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"color": gComputedStyle.color};
      testTextAttrs(ID, 50, attrs, defAttrs, 50, 57);

      tempElem = tempElem.firstChild.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"font-weight": gComputedStyle.fontWeight,
               "color": gComputedStyle.color};
      testTextAttrs(ID, 57, attrs, defAttrs, 57, 61);

      tempElem = tempElem.parentNode;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"color": gComputedStyle.color};
      testTextAttrs(ID, 61, attrs, defAttrs, 61, 68);

      //////////////////////////////////////////////////////////////////////////
      // area9, different single style spans in styled paragraph
      ID = "area9";
      tempElem = document.getElementById(ID);
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      defAttrs = {
        "font-style": gComputedStyle.fontStyle,
        "font-size": "10pt",
        "background-color": "rgb(255, 255, 255)",
        "font-weight": gComputedStyle.fontWeight,
        "color": gComputedStyle.color,
        "font-family": gComputedStyle.fontFamily,
        "text-position": gComputedStyle.verticalAlign
      };

      testDefaultTextAttrs(ID, defAttrs);

      attrs = {};
      testTextAttrs(ID, 0, attrs, defAttrs, 0, 6);

      attrs = { "font-size": "12pt" };
      testTextAttrs(ID, 7, attrs, defAttrs, 6, 12);

      attrs = {};
      testTextAttrs(ID, 13, attrs, defAttrs, 12, 21);

      // Walk to the span with the different background color
      tempElem = tempElem.firstChild.nextSibling.nextSibling.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = { "background-color": gComputedStyle.backgroundColor };
      testTextAttrs(ID, 22, attrs, defAttrs, 21, 36);

      attrs = {};
      testTextAttrs(ID, 37, attrs, defAttrs, 36, 44);

      // Walk from the background color span to the one with font-style
      tempElem = tempElem.nextSibling.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = { "font-style": gComputedStyle.fontStyle };
      testTextAttrs(ID, 45, attrs, defAttrs, 44, 61);

      attrs = {};
      testTextAttrs(ID, 62, attrs, defAttrs, 61, 69);

      // Walk from span with font-style to the one with font-family.
      tempElem = tempElem.nextSibling.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = { "font-family": gComputedStyle.fontFamily };
      testTextAttrs(ID, 70, attrs, defAttrs, 69, 83);

      attrs = {};
      testTextAttrs(ID, 84, attrs, defAttrs, 83, 91);

      attrs = { "text-underline-style": "solid" };
      testTextAttrs(ID, 92, attrs, defAttrs, 91, 101);

      attrs = {};
      testTextAttrs(ID, 102, attrs, defAttrs, 101, 109);

      attrs = { "text-line-through-style": "solid" };
      testTextAttrs(ID, 110, attrs, defAttrs, 109, 122);

      attrs = {};
      testTextAttrs(ID, 123, attrs, defAttrs, 122, 130);

      // area10, different single style spans in non-styled paragraph
      ID = "area10";
      tempElem = document.getElementById(ID);
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      defAttrs = {
        "font-style": gComputedStyle.fontStyle,
        "font-size": "12pt",
        "background-color": "rgb(255, 255, 255)",
        "font-weight": gComputedStyle.fontWeight,
        "color": gComputedStyle.color,
        "font-family": gComputedStyle.fontFamily,
        "text-position": gComputedStyle.verticalAlign
      };

      testDefaultTextAttrs(ID, defAttrs);

      attrs = {};
      testTextAttrs(ID, 0, attrs, defAttrs, 0, 7);

      attrs = { "font-size": "14pt" };
      testTextAttrs(ID, 7, attrs, defAttrs, 7, 13);

      attrs = {};
      testTextAttrs(ID, 13, attrs, defAttrs, 13, 22);

      // Walk to the span with the different background color
      tempElem = tempElem.firstChild.nextSibling.nextSibling.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = { "background-color": gComputedStyle.backgroundColor };
      testTextAttrs(ID, 23, attrs, defAttrs, 22, 37);

      attrs = {};
      testTextAttrs(ID, 38, attrs, defAttrs, 37, 45);

      // Walk from the background color span to the one with font-style
      tempElem = tempElem.nextSibling.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"font-style": gComputedStyle.fontStyle};
      testTextAttrs(ID, 46, attrs, defAttrs, 45, 62);

      attrs = {};
      testTextAttrs(ID, 63, attrs, defAttrs, 62, 70);

      // Walk from span with font-style to the one with font-family.
      tempElem = tempElem.nextSibling.nextSibling;
      gComputedStyle = document.defaultView.getComputedStyle(tempElem, "");
      attrs = {"font-family": gComputedStyle.fontFamily};
      testTextAttrs(ID, 71, attrs, defAttrs, 70, 84);

      attrs = {};
      testTextAttrs(ID, 85, attrs, defAttrs, 84, 92);

      attrs = { "text-underline-style": "solid" };
      testTextAttrs(ID, 93, attrs, defAttrs, 92, 102);

      attrs = {};
      testTextAttrs(ID, 103, attrs, defAttrs, 102, 110);

      attrs = { "text-line-through-style": "solid" };
      testTextAttrs(ID, 111, attrs, defAttrs, 110, 123);

      attrs = {};
      testTextAttrs(ID, 124, attrs, defAttrs, 123, 131);

      //////////////////////////////////////////////////////////////////////////
      // test spelling text attributes
      testSpellTextAttrs(); // Will call SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body style="font-size: 12pt">

  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=345759"
     title="Implement text attributes">
    Mozilla Bug 345759
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <p id="area1" style="font-size: smaller">Normal <b>Bold</b> Normal</p>
  <p id="area2" style="font-size: 120%">Normal <b>Bold <i>Italic </i>Bold</b> Normal</p>
  <p id="area3" style="background-color: blue;">
    <span style="color: green; background-color: rgb(0, 0, 255)">
      Green
      <span style="color: red">but children are red</span>
    </span><span style="color: green; background-color: rgb(255, 255, 0);">
      Another green section.
    </span>
  </p>
  <p id="area4">
    <span style="color: green">
      Green
    </span><span style="color: green">
      Green too
      <span style="color: red">with red children</span>
      Green again
    </span>
  </p>
  <p id="area5">
    <span style="color: green">Green</span>
    <img src="moz.png" alt="image"/>
    <span style="color: red">Red</span>Normal
  </p>
  <p id="area6">
    This <sup>sentence</sup> has the word 
    <span style="vertical-align:super;">sentence</span> in 
    <sub>superscript</sub> and 
    <span style="vertical-align:sub;">subscript</span>
  </p>

  <p lang="en" id="area7">
    <span lang="ru">Привет</span>
    <span style="background-color: blue">Blue BG color</span>
    <span lang="de">Ich bin/Du bist</span>
    <span lang="en">
      Normal
      <span style="color: magenta">Magenta<b>Bold</b>Magenta</span>
    </span>
  </p>

  <input id="area8"/>

  <p id="area9" style="font-size: smaller">Small
    <span style="font-size: 120%">bigger</span> smaller
    <span style="background-color: blue;">background blue</span> normal
    <span style="font-style: italic;">Different styling</span> normal
    <span style="font-family: tahoma;">Different font</span> normal
    <span style="text-decoration: underline;">underlined</span> normal
    <span style="text-decoration: line-through;">strikethrough</span> normal
  </p>

  <p id="area10">Normal
    <span style="font-size: 120%">bigger</span> smaller
    <span style="background-color: blue;">background blue</span> normal
    <span style="font-style: italic;">Different styling</span> normal
    <span style="font-family: tahoma;">Different font</span> normal
    <span style="text-decoration: underline;">underlined</span> normal
    <span style="text-decoration: line-through;">strikethrough</span> normal
  </p>
</body>
</html>
